RSS-FINAL-4_GraphQL-functionality-fix (#22)
* fix: RSS-FINAL-4_GraphQL-functionality-fix fix GraphQl Page

* feat: RSS-FINAL-4_GraphQL-functionality-fix add metadata  for page GraphQL

* feat: add new component

* fix: RSS-FINAL-4_GraphQL-functionality-fix fix prettify

* fix: RSS-FINAL-4_GraphQL-functionality-fix fix props

---------
