+++ graphiql-app/src/app/[locale]/graphql/[...slug]/types.ts
@@ -7,5 +7,9 @@ interface GraphQLPageProps {
   params: GraphQLPageParams;
   searchParams: Record<string, string>;
 }
+interface GraphQLPageClientProps extends GraphQLPageProps {
+  url: string;
+  bodyBase64: string;
+}
 
-export type { GraphQLPageParams, GraphQLPageProps };
+export type { GraphQLPageParams, GraphQLPageProps, GraphQLPageClientProps };
