+++ graphiql-app/src/app/[locale]/graphql/page.tsx
@@ -1,10 +1,17 @@
 import { Box, Button, Container, Link, Typography } from '@mui/material';
 import { useTranslations } from 'next-intl';
 import BarChartRoundedIcon from '@mui/icons-material/BarChartRounded';
+import { Metadata } from 'next';
+
+export const metadata: Metadata = {
+  title: 'GraphQL',
+  description: 'Jump page for first request using GraphQL',
+  authors: [{ name: 'Oleksandr', url: 'https://github.com/aleks6699' }],
+  keywords: ['GraphQL', 'Playground', 'API', 'QueryHub', 'Next.js'],
+};
 
 export default function GraphQLPage() {
   const locale = useTranslations();
-
   return (
     <Container
       sx={{
