+++ graphiql-app/src/components/PrettyButton/types.ts
@@ -1,0 +1,7 @@
+interface PrettyButtonProps {
+  content: string;
+  isQuery: boolean;
+  onChange: (value: string) => void;
+}
+
+export default PrettyButtonProps;
