+++ graphiql-app/src/utils/prettify.tsx
